anstyle
Struct Effects

pub struct Effects(/* private fields */);

A set of text effects

Examples

let effects = anstyle::Effects::BOLD | anstyle::Effects::UNDERLINE;

Implementations

impl Effects

pub const BOLD: Self

pub const DIMMED: Self

pub const ITALIC: Self

Not widely supported. Sometimes treated as inverse or blink

pub const UNDERLINE: Self

Style extensions exist for Kitty, VTE, mintty and iTerm2.

pub const DOUBLE_UNDERLINE: Self

pub const CURLY_UNDERLINE: Self

pub const DOTTED_UNDERLINE: Self

pub const DASHED_UNDERLINE: Self

pub const BLINK: Self

pub const INVERT: Self

Swap foreground and background colors; inconsistent emulation

pub const HIDDEN: Self

pub const STRIKETHROUGH: Self

Characters legible but marked as if for deletion. Not supported in Terminal.app

pub const fn new() -> Self

No effects enabled

Examples

let effects = anstyle::Effects::new();

pub const fn is_plain(self) -> bool

Check if no effects are enabled

Examples

let effects = anstyle::Effects::new();
assert!(effects.is_plain());

let effects = anstyle::Effects::BOLD | anstyle::Effects::UNDERLINE;
assert!(!effects.is_plain());

pub const fn contains(self, other: Effects) -> bool

Returns true if all of the effects in other are contained within self.

Examples

let effects = anstyle::Effects::BOLD | anstyle::Effects::UNDERLINE;
assert!(effects.contains(anstyle::Effects::BOLD));

let effects = anstyle::Effects::new();
assert!(!effects.contains(anstyle::Effects::BOLD));

pub const fn insert(self, other: Effects) -> Self

Inserts the specified effects in-place.

Examples

let effects = anstyle::Effects::new().insert(anstyle::Effects::new());
assert!(effects.is_plain());

let effects = anstyle::Effects::new().insert(anstyle::Effects::BOLD);
assert!(effects.contains(anstyle::Effects::BOLD));

pub const fn remove(self, other: Effects) -> Self

Removes the specified effects in-place.

Examples

let effects = (anstyle::Effects::BOLD | anstyle::Effects::UNDERLINE).remove(anstyle::Effects::BOLD);
assert!(!effects.contains(anstyle::Effects::BOLD));
assert!(effects.contains(anstyle::Effects::UNDERLINE));

pub const fn clear(self) -> Self

Reset all effects in-place

let effects = (anstyle::Effects::BOLD | anstyle::Effects::UNDERLINE).clear();
assert!(!effects.contains(anstyle::Effects::BOLD));
assert!(!effects.contains(anstyle::Effects::UNDERLINE));

pub const fn set(self, other: Self, enable: bool) -> Self

Enable or disable the specified effects depending on the passed value.

Examples

let effects = anstyle::Effects::new().set(anstyle::Effects::BOLD, true);
assert!(effects.contains(anstyle::Effects::BOLD));

pub fn iter(self) -> EffectIter

Iterate over enabled effects

pub fn render(self) -> impl Display + Copy

Render the ANSI code