anstyle
Struct Style

pub struct Style { /* private fields */ }

ANSI Text styling

You can print a Style to render the corresponding ANSI code. Using the alternate flag # will render the ANSI reset code, if needed. Together, this makes it convenient to render styles using inline format arguments.

Examples

let style = anstyle::Style::new().bold();

let value = 42;
println!("{style}{value}{style:#}");

Implementations

impl Style

Core

pub const fn new() -> Self

No effects enabled

Examples

let style = anstyle::Style::new();

pub const fn fg_color(self, fg: Option<Color>) -> Self

Set foreground color

Examples

let style = anstyle::Style::new().fg_color(Some(anstyle::AnsiColor::Red.into()));

pub const fn bg_color(self, bg: Option<Color>) -> Self

Set background color

Examples

let style = anstyle::Style::new().bg_color(Some(anstyle::AnsiColor::Red.into()));

pub const fn underline_color(self, underline: Option<Color>) -> Self

Set underline color

Examples

let style = anstyle::Style::new().underline_color(Some(anstyle::AnsiColor::Red.into()));

pub const fn effects(self, effects: Effects) -> Self

Set text effects

Examples

let style = anstyle::Style::new().effects(anstyle::Effects::BOLD | anstyle::Effects::UNDERLINE);

pub fn render(self) -> impl Display + Copy

Render the ANSI code

Style also implements Display directly, so calling this method is optional.

pub fn write_to(self, write: &mut dyn Write) -> Result<()>

Write the ANSI code

pub fn render_reset(self) -> impl Display + Copy

Renders the relevant Reset code

Unlike Reset::render, this will elide the code if there is nothing to reset.

pub fn write_reset_to(self, write: &mut dyn Write) -> Result<()>

Write the relevant Reset code

Unlike Reset::render, this will elide the code if there is nothing to reset.

impl Style

Convenience

pub const fn bold(self) -> Self

Apply bold effect

Examples

let style = anstyle::Style::new().bold();

pub const fn dimmed(self) -> Self

Apply dimmed effect

Examples

let style = anstyle::Style::new().dimmed();

pub const fn italic(self) -> Self

Apply italic effect

Examples

let style = anstyle::Style::new().italic();

pub const fn underline(self) -> Self

Apply underline effect

Examples

let style = anstyle::Style::new().underline();

pub const fn blink(self) -> Self

Apply blink effect

Examples

let style = anstyle::Style::new().blink();

pub const fn invert(self) -> Self

Apply invert effect

Examples

let style = anstyle::Style::new().invert();

pub const fn hidden(self) -> Self

Apply hidden effect

Examples

let style = anstyle::Style::new().hidden();

pub const fn strikethrough(self) -> Self

Apply strikethrough effect

Examples

let style = anstyle::Style::new().strikethrough();

impl Style

Reflection

pub const fn get_fg_color(self) -> Option<Color>

Get the foreground color

pub const fn get_bg_color(self) -> Option<Color>

Get the background color

pub const fn get_underline_color(self) -> Option<Color>

pub const fn get_effects(self) -> Effects

pub const fn is_plain(self) -> bool

Check if no styling is enabled